+++ manufacturing-ai-suite/industrial-edge-insights-vision/tests/robot_files/test_main_industrial_edge_insights_vision.robot
@@ -1,12 +1,12 @@
 ***Settings***
 Documentation    This is main test case file.
-Library          test_suit_spineapps_cases.py
-
-***Variables***
-${IS_OPEN_EDGE}    false
-
+Library          test_suit.py
 
 ***Keywords***
+
+# NOTE: The warning means that test_suit.py does not define any Robot Framework keywords.
+# To fix this, ensure that test_suit.py contains functions decorated with @keyword from robot.api.deco,
+# or that you use the Python class-based library style with public methods.
 
 App_Test_case_001
     [Documentation]     Verify Happy Path for PDD - CPU
@@ -35,5 +35,5 @@ APP_TC_001
 APP_TC_002
     [Documentation]    Verify Happy Path for WELD - CPU
     [Tags]      app
-    ${Status}    Run Keyword And Return Status   APp_Test_case_002
+    ${Status}    Run Keyword And Return Status   App_Test_case_002
     Should Not Be Equal As Integers    ${Status}    0
