+++ manufacturing-ai-suite/industrial-edge-insights-vision/tests/robot_files/test_suit.py
@@ -5,20 +5,20 @@ import os
 env = os.environ.copy()
 
 
-class test_suit_spineapps_cases(unittest.TestCase):
+class test_suit(unittest.TestCase):
 
     ##################################################################################################################################################
-    #                                   Test case with spineapps Use case Only
+    #                                   Test case with industrial_edge_insights_vision apps use cases
     ##################################################################################################################################################
     
     def TC_001_app(self):
         env["TEST_CASE"] = "APP001"
-        ret = subprocess.call("nosetests3 --nocapture -v ../functional_tests/spineapps.py:test_spineapps_cases.test_spineapps", shell=True, env=env)
+        ret = subprocess.call("nosetests3 --nocapture -v ../functional_tests/apps.py:test_apps_cases.test_apps", shell=True, env=env)
         return ret
     
     def TC_002_app(self):
         env["TEST_CASE"] = "APP002"
-        ret = subprocess.call("nosetests3 --nocapture -v ../functional_tests/spineapps.py:test_spineapps_cases.test_spineapps", shell=True, env=env)
+        ret = subprocess.call("nosetests3 --nocapture -v ../functional_tests/apps.py:test_apps_cases.test_apps", shell=True, env=env)
         return ret
